Use floatValue and doubleValue to validate values
We don't need intermediate conversion to the BigInteger
as it can lead to heavy computations of pow(BigInteger.TEN, big_exponent)
causing queries to hang.

(cherry picked from commit 3481af304449c3b306d769af3f7f2e40f58d3170)

+++ server/src/main/java/io/crate/types/DoubleType.java
@@ -23,7 +23,6 @@ package io.crate.types;
 
 import java.io.IOException;
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.util.List;
 import java.util.function.Function;
 
@@ -119,8 +118,8 @@ public class DoubleType extends DataType<Double> implements FixedWidthType, Stre
         }
     };
 
-    private static final BigInteger DOUBLE_MAX = BigDecimal.valueOf(Double.MAX_VALUE).toBigInteger();
-    private static final BigInteger DOUBLE_MIN = BigDecimal.valueOf(-Double.MAX_VALUE).toBigInteger();
+    private static final BigDecimal DOUBLE_MAX = BigDecimal.valueOf(Double.MAX_VALUE);
+    private static final BigDecimal DOUBLE_MIN = BigDecimal.valueOf(-Double.MAX_VALUE);
 
     private DoubleType() {
     }
@@ -159,8 +158,12 @@ public class DoubleType extends DataType<Double> implements FixedWidthType, Stre
         } else if (value instanceof String s) {
             return Double.valueOf(s);
         } else if (value instanceof BigDecimal bigDecimalValue) {
-            if (DOUBLE_MAX.compareTo(bigDecimalValue.toBigInteger()) <= 0
-                || DOUBLE_MIN.compareTo(bigDecimalValue.toBigInteger()) >= 0) {
+            // Preferring compareTo over equals as it gives safer equality check.
+            // From the compareTo docs:
+            // Two BigDecimal objects that are equal in value but have a different scale (like 2.0 and 2.00)
+            // are considered equal by this method.
+            if (DOUBLE_MAX.compareTo(bigDecimalValue) <= 0
+                || DOUBLE_MIN.compareTo(bigDecimalValue) >= 0) {
                 throw new IllegalArgumentException(getName() + " value out of range: " + value);
             }
             return bigDecimalValue.doubleValue();
